zn - finish all the refactor, and replace the hello with earthquakes
