+++ pom.xml
@@ -223,7 +223,7 @@
                     <archive>
                         <manifest>
                             <!-- full package name of class with the main you want to run -->
-			    <mainClass>hello.Application</mainClass>
+			    <mainClass>earthquakes.Application</mainClass>
 
                         </manifest>
                     </archive>
