+++ src/main/java/earthquakes/AuthControllerAdvice.java
@@ -1,14 +1,14 @@
-package hello;
+package earthquakes;
 
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
 import org.springframework.security.oauth2.core.user.OAuth2User;
 import org.springframework.web.bind.annotation.ControllerAdvice;
-import hello.repositories.UserRepository;
-import hello.entities.AppUser;
+import earthquakes.repositories.UserRepository;
+import earthquakes.entities.AppUser;
 import java.util.List;
-import hello.services.MembershipService;
+import earthquakes.services.MembershipService;
 
 @ControllerAdvice
 public class AuthControllerAdvice {
