+++ src/main/java/earthquakes/UserRepository.java
@@ -1,9 +1,9 @@
-package hello.repositories;
+package earthquakes.repositories;
 
 import java.util.List;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import hello.entities.AppUser;
+import earthquakes.entities.AppUser;
 
 @Repository
 public interface UserRepository extends CrudRepository<AppUser, Long> {
