+++ src/main/java/earthquakes/controllers/EarthquakesController.java
@@ -1,6 +1,6 @@
-package hello.controllers;
+package earthquakes.controllers;
 
-import hello.services.EarthquakeQueryService;
+import earthquakes.services.EarthquakeQueryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.oauth2.client.authentication.OAuth2AuthenticationToken;
 import org.springframework.security.oauth2.client.registration.ClientRegistration;
@@ -13,8 +13,8 @@ import java.util.Map;
 import java.util.HashMap;
 
 import com.nimbusds.oauth2.sdk.client.ClientReadRequest;
-import hello.geojson.FeatureCollection;
-import hello.searches.EqSearch;
+import earthquakes.geojson.FeatureCollection;
+import earthquakes.searches.EqSearch;
 
 @Controller
 public class EarthquakesController {
