+++ src/main/java/earthquakes/searches/EqSearch.java
@@ -1,4 +1,4 @@
-package hello.searches;
+package earthquakes.searches;
 
 public class EqSearch{
     private int distance;
