+++ src/main/java/earthquakes/services/GithubOrgMembershipService.java
@@ -1,4 +1,4 @@
-package hello.services;
+package earthquakes.services;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
