+++ src/main/resources/application.properties
@@ -1,5 +1,5 @@
 server.port=${PORT:8080}
-logging.level.hello=INFO
+logging.level.earthquakes=INFO
 spring.security.oauth2.client.registration.github.scope: user,read:org
 app_github_org: ucsb-cs56-f19
