+++ src/test/java/hello/BootstrapLiterals.java
@@ -1,4 +1,4 @@
-package hello;
+package earthquakes;
 
 /** 
  * Literals for testing whether Bootstrap is loaded correctly
