+++ src/test/java/hello/EarthquakeSearchTest.java
@@ -1,4 +1,4 @@
-package hello;
+package earthquakes;
 
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
@@ -22,7 +22,7 @@ import org.junit.Before;
 
 import static org.springframework.security.test.web.servlet.request.SecurityMockMvcRequestPostProcessors.authentication;
 
-import hello.controllers.EarthquakesController;
+import earthquakes.controllers.EarthquakesController;
 
 @RunWith(SpringRunner.class)
 @WebMvcTest(EarthquakesController.class)
