+++ src/test/java/hello/HomePageTest.java
@@ -1,4 +1,4 @@
-package hello;
+package earthquakes;
 
 // import static org.hamcrest.Matchers.equalTo;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
@@ -17,7 +17,7 @@ import org.springframework.security.oauth2.client.registration.ClientRegistratio
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.xpath;
-import hello.controllers.HomeController;
+import earthquakes.controllers.HomeController;
 
 
 @RunWith(SpringRunner.class)
